refactor: split eval AOT analysis
